Sanitize function-to-CU allocations
