+++ src/deviceconfig_validation.cpp
@@ -97,30 +97,3 @@ SuccessResultWithMessage DeviceConfig::ValidateRuntimeConfig(const RuntimeConfig
 
     return { true, "" };
 }
-
-SuccessResultWithMessage DeviceConfig::SetRuntimeConfig(const RuntimeConfig& config, bool skipWrite)
-{
-    auto [isValid, validationMessage] = ValidateRuntimeConfig(config);
-    if (!isValid)
-        return { false, validationMessage };
-
-    const bool changed =
-        runtimeTopology.width != config.topology.width
-        || runtimeTopology.height != config.topology.height
-        || runtimeTopology.serpentine != config.topology.serpentine
-        || runtimeOutputs.driver != config.outputs.driver
-        || runtimeOutputs.channelCount != config.outputs.channelCount
-        || runtimeOutputs.outputPins != config.outputs.outputPins
-        || runtimeOutputs.colorOrder != config.outputs.colorOrder;
-
-    runtimeTopology = config.topology;
-    runtimeOutputs = config.outputs;
-
-    if (!skipWrite)
-        SaveToJSON();
-
-    if (changed && !skipWrite)
-        LogRuntimeConfig("runtime config changed");
-
-    return { true, "" };
-}
