+++ src/effectmanager_access.cpp
@@ -72,6 +72,11 @@ void EffectManager::AddEffectEventListener(IEffectEventListener& listener)
     _effectEventListeners.emplace_back(listener);
 }
 
+std::vector<std::shared_ptr<GFXBase>> & EffectManager::GetBaseGraphics()
+{
+    return _gfx;
+}
+
 // Must provide at least one drawing instance, like the first matrix or strip we are drawing on
 GFXBase& EffectManager::g(int iChannel)
 {
@@ -92,31 +97,6 @@ bool EffectManager::IsEffectEnabled(size_t i) const
         return false;
     }
     return _vEffects[i]->IsEnabled();
-}
-
-void EffectManager::PlayAll(bool bPlayAll)
-{
-    std::lock_guard effectGuard(g_effect_manager_mutex);
-    _bPlayAll = bPlayAll;
-}
-
-void EffectManager::SetInterval(uint interval, bool skipSave)
-{
-    std::lock_guard effectGuard(g_effect_manager_mutex);
-
-    // Reject/ignore intervals smaller than a second, but allow 0 (infinity)
-    if (interval > 0 && interval < 1000)
-        return;
-
-    _effectInterval = interval;
-
-    if (!skipSave)
-        SaveEffectManagerConfig();
-
-    {
-        std::lock_guard listenerGuard(_listenerMutex);
-        INFORM_EVENT_LISTENERS(_effectEventListeners, IEffectEventListener::OnIntervalChanged, interval);
-    }
 }
 
 std::vector<std::shared_ptr<LEDStripEffect>> EffectManager::EffectsList() const
