+++ src/gfxbase_transforms.cpp
@@ -43,6 +43,77 @@ void GFXBase::ResetOscillators()
     std::fill_n(p, 6, 0);
 }
 
+void GFXBase::MoveInwardX(int startY, int endY)
+{
+    const int width = static_cast<int>(_width);
+    const int height = static_cast<int>(_height);
+    const int halfWidth = width / 2;
+    if (leds == nullptr || halfWidth <= 1 || height <= 0)
+        return;
+
+    startY = std::max(0, startY);
+    endY = std::min(height - 1, endY);
+
+    for (int y = startY; y <= endY; y++)
+    {
+        // Keep both halves inside the row. The previous loops read x+1 at
+        // width, which can corrupt unrelated pixels or memory at row edges.
+        for (int x = halfWidth - 1; x > 0; x--)
+            leds[XY(x, y)] = leds[XY(x - 1, y)];
+
+        for (int x = halfWidth; x < width - 1; x++)
+            leds[XY(x, y)] = leds[XY(x + 1, y)];
+    }
+}
+
+void GFXBase::MoveOutwardsX(int startY, int endY)
+{
+    const int width = static_cast<int>(_width);
+    const int height = static_cast<int>(_height);
+    const int halfWidth = width / 2;
+    if (leds == nullptr || halfWidth <= 1 || height <= 0)
+        return;
+
+    startY = std::max(0, startY);
+    endY = std::min(height - 1, endY);
+
+    for (int y = startY; y <= endY; y++)
+    {
+        for (int x = 0; x < halfWidth - 1; x++)
+            leds[XY(x, y)] = leds[XY(x + 1, y)];
+
+        for (int x = width - 1; x > halfWidth; x--)
+            leds[XY(x, y)] = leds[XY(x - 1, y)];
+    }
+}
+
+void GFXBase::MoveX(uint8_t delta) const
+{
+    for (int y = 0; y < (int)_height; y++)
+    {
+        for (int x = 0; x < (int)_width - delta; x++)
+            leds[XY(x, y)] = leds[XY(x + delta, y)];
+        for (int x = (int)_width - delta; x < (int)_width; x++)
+            leds[XY(x, y)] = leds[XY(x + delta - (int)_width, y)];
+    }
+}
+
+void GFXBase::MoveY(uint8_t delta) const
+{
+    CRGB tmp = 0;
+    for (int x = 0; x < (int)_width; x++)
+    {
+        tmp = leds[XY(x, 0)];
+        for (int m = 0; m < delta; m++)
+        {
+            for (int y = 0; y < (int)_height - 1; y++)
+                leds[XY(x, y)] = leds[XY(x, y + 1)];
+
+            leds[XY(x, (int)_height - 1)] = tmp;
+        }
+    }
+}
+
 // All the Caleidoscope functions work directly within the screenbuffer (leds array).
 // Draw whatever you like in the area x(0-15) and y (0-15) and then copy it around.
